se soluciono errores en test incremental (series)

+++ database/migrations/2017_04_28_231653_create_series_table.php
@@ -12,17 +12,17 @@ class CreateSeriesTable extends Migration
         Schema::create('series', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('cantidad_series');
-            $table->integer('peso_corporal');
-            $table->integer('peso_externo');
+            $table->float('peso_corporal');
+            $table->float('peso_externo');
 
-            $table->integer('masa');//Aca va Masa, que es una suma de peso externo con peso corporal
+            $table->float('masa');//Aca va Masa, que es una suma de peso externo con peso corporal
 
-            $table->integer('potencia_impulsiva');
+            $table->float('potencia_impulsiva');
 
             $table->float('potencia_relativa');//Acá va potencia Relativa = P.Impulsiva/Peso Corporal
 
-            $table->integer('velocidad_impulsiva');
-            $table->integer('fuerza_impulsiva');
+            $table->float('velocidad_impulsiva');
+            $table->float('fuerza_impulsiva');
 
             $table->integer('cantidad_repeticiones');
             $table->integer('mejor_serie');
